Projeto BackgroundRemover/requirements.txt
rembg==2.0.60
onnxruntime==1.20.1
pillow==11.0.0